refactor : AutoLotto 클래스 선언할 때 한 번만 1부터 45까지 숫자 모음 만들기
